fix: no focusable elements cases
