+++ src/Pages/Shared/LinkedCIDetailsModal/LinkedCIDetailsModal.tsx
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { useEffect, useMemo, useRef } from 'react'
+import { useMemo, useRef } from 'react'
 import { useParams } from 'react-router-dom'
 
 import {
@@ -36,7 +36,6 @@ import {
 } from '@devtron-labs/devtron-fe-common-lib'
 
 import { ReactComponent as Close } from '../../../assets/icons/ic-close.svg'
-import { preventBodyScroll } from '../../../components/common'
 import { API_STATUS_CODES, SELECT_ALL_VALUE } from '../../../config'
 import { ALL_ENVIRONMENT_OPTION, SortableKeys } from './constants'
 import LinkedCIAppList from './LinkedCIAppList'
@@ -112,14 +111,6 @@ const LinkedCIDetailsModal = ({ handleClose, workflows }: LinkedCIDetailModalPro
     const selectOptions = [ALL_ENVIRONMENT_OPTION, ...(envList ?? []).map((env) => ({ label: env, value: env }))]
 
     const selectedOption = selectOptions.find((option) => option.value === environment) ?? ALL_ENVIRONMENT_OPTION
-
-    useEffect(() => {
-        preventBodyScroll(true)
-
-        return () => {
-            preventBodyScroll(false)
-        }
-    }, [])
 
     const showLoadingState = loading || getIsRequestAborted(error)
