+++ src/components/app/details/appDetails/AppDetailsCDModal.tsx
@@ -70,7 +70,7 @@ const AppDetailsCDModal = ({
 
     const renderCDModal = () =>
         (mode === URL_PARAM_MODE_TYPE.LIST || mode === URL_PARAM_MODE_TYPE.REVIEW_CONFIG) && (
-            <VisibleModal className="" parentClassName="dc__overflow-hidden" close={closeCDModal}>
+            <VisibleModal parentClassName="dc__overflow-hidden" close={closeCDModal}>
                 <div className="modal-body--cd-material h-100 contains-diff-view flexbox-col" onClick={stopPropagation}>
                     <CDMaterial
                         materialType={materialType}
