+++ src/components/app/details/triggerView/TriggerView.tsx
@@ -52,7 +52,6 @@ import {
     getCIPipelineURL,
     importComponentFromFELibrary,
     InValidHostUrlWarningBlock,
-    preventBodyScroll,
     sortObjectArrayAlphabetically,
     withAppContext,
 } from '../../../common'
@@ -525,7 +524,6 @@ class TriggerView extends Component<TriggerViewProps, TriggerViewState> {
                 },
                 () => {
                     this.getWorkflowStatus()
-                    preventBodyScroll(true)
                 },
             )
         })
@@ -681,7 +679,6 @@ class TriggerView extends Component<TriggerViewProps, TriggerViewState> {
                 },
                 () => {
                     this.getWorkflowStatus()
-                    preventBodyScroll(true)
                 },
             )
         })
@@ -772,7 +769,6 @@ class TriggerView extends Component<TriggerViewProps, TriggerViewState> {
             cdNodeId,
             nodeType,
         })
-        preventBodyScroll(true)
 
         const newParams = new URLSearchParams(this.props.location.search)
         newParams.set(
@@ -821,7 +817,6 @@ class TriggerView extends Component<TriggerViewProps, TriggerViewState> {
                 nodeType: 'CD',
             },
             () => {
-                preventBodyScroll(true)
                 this.getWorkflowStatus()
             },
         )
@@ -924,7 +919,6 @@ class TriggerView extends Component<TriggerViewProps, TriggerViewState> {
                             invalidateCache: false,
                         },
                         () => {
-                            preventBodyScroll(false)
                             this.getWorkflowStatus()
                             if (this.props.isJobView) {
                                 this.getWorkflows()
@@ -1063,7 +1057,6 @@ class TriggerView extends Component<TriggerViewProps, TriggerViewState> {
     }
 
     closeCIModal = (): void => {
-        preventBodyScroll(false)
         this.abortController.abort()
         this.setState({ showMaterialRegexModal: false })
         this.props.history.push(this.props.match.url)
@@ -1075,17 +1068,14 @@ class TriggerView extends Component<TriggerViewProps, TriggerViewState> {
         this.props.history.push({
             search: '',
         })
-        preventBodyScroll(false)
         this.getWorkflowStatus()
     }
 
     closeApprovalModal = (e: React.MouseEvent): void => {
         e.stopPropagation()
-        preventBodyScroll(false)
         this.props.history.push({
             search: '',
         })
-        preventBodyScroll(false)
         this.getWorkflowStatus()
     }
 
@@ -1303,7 +1293,7 @@ class TriggerView extends Component<TriggerViewProps, TriggerViewState> {
             const material = cdNode[this.state.materialType] || []
 
             return (
-                <VisibleModal className="" parentClassName="dc__overflow-hidden" close={this.closeCDModal}>
+                <VisibleModal parentClassName="dc__overflow-hidden" close={this.closeCDModal}>
                     <div
                         className={`modal-body--cd-material h-100 flexbox-col contains-diff-view ${
                             material.length > 0 ? '' : 'no-material'
