+++ src/components/security/SecurityScansTab/utils.tsx
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { Severity } from '@devtron-labs/devtron-fe-common-lib'
+import { Badge, SeveritiesDTO, Severity, SeverityCount } from '@devtron-labs/devtron-fe-common-lib'
 
 import { SearchType, SecurityScansTabMultiFilterKeys, SecurityScansTabSingleFilterKeys, SeverityMapping } from './types'
 
@@ -45,4 +45,32 @@ export const getSeverityFilterLabelFromValue = (severity: string) => {
         default:
             return SeverityMapping.unknown
     }
+}
+
+const SEVERITY_ORDER = [
+    { key: SeveritiesDTO.CRITICAL, label: 'Critical', variant: 'negative' },
+    { key: SeveritiesDTO.HIGH, label: 'High', variant: 'custom', fontColor: 'R500', bgColor: 'R100' },
+    { key: SeveritiesDTO.MEDIUM, label: 'Medium', variant: 'custom', fontColor: 'O600', bgColor: 'O100' },
+    { key: SeveritiesDTO.LOW, label: 'Low', variant: 'warning' },
+    { key: SeveritiesDTO.UNKNOWN, label: 'Unknown', variant: 'neutral' },
+] as const
+
+export const getSeverityWithCount = (severityCount: SeverityCount) => {
+    // eslint-disable-next-line no-restricted-syntax
+    for (const item of SEVERITY_ORDER) {
+        if (severityCount[item.key]) {
+            if (item.variant === 'custom') {
+                return (
+                    <Badge
+                        label={`${severityCount[item.key]} ${item.label}`}
+                        variant="custom"
+                        fontColor={item.fontColor}
+                        bgColor={item.bgColor}
+                    />
+                )
+            }
+            return <Badge label={`${severityCount[item.key]} ${item.label}`} variant={item.variant} />
+        }
+    }
+    return <Badge label="Passed" variant="positive" />
 }
